Fix site generation related issues
Previously, we were splitting and concatenating files a couple of
times to build the relaton XML links, and recently we have added
a dedicated split interface to optimize it, and which does all of
the work. But it did introduced some issues, it wasn't generating
the site properly.

This commit fixes these issues, so we can use this new interface
and also keep all existing functionality intact.