Fix relaton link generation issues
We've added some new changes to resolve the relaton xml link
in relaton-cli `v0.1.10`, but that was never pushed to rubygems.

So, when this application was switched to use the official gem
then it was using the old version, and that was causing the issue
with the relaton xml link.

Now we've updated the relaton-cli, and it also specify minimum
version in this repo, so it should fix this issue now.

Fixes 27